Redo how fastmath functions are implemented
Previously, each fast math function was implemented by having an
corresponding `ops::fast_X<T>` functor for each `ops::X<T>` functor.

This commit rewrite this logic to instead have a special
`apply_fastmath_impl<...>` helper that falls back to `apply_impl<...>`
if there is not fast math version available. This means the normal
apply logic is used if there is not fast math version, which was
previously not the case.

Additionally, this commit adds a global `KERNEL_FLOAT_FAST_MATH`
define that can be used to turn on fast math mode.